+++ examples/ephemeral-resources/porkbun_ssl/ephemeral-resource.tf
@@ -1,0 +1,3 @@
+ephemeral "porkbun_ssl" "example" {
+  domain = "example.com"
+}